+++ js/onlyoffice-docspace-formater.js
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -17,66 +17,84 @@
 */
 
 (function (Drupal) {
+  const defaultConfig = {
+    width: "100%",
+    height: "100%",
+    locale: drupalSettings.OODSP_Settings.locale,
+  };
+
   DocspaceIntegrationSdk.initScript("oodsp-api-js", drupalSettings.OODSP_Settings.url).then(
     function() {
+      const firstElement = Object.values(drupalSettings.OODSPFormatterData)[0];
+
       for (var frameId in drupalSettings.OODSPFormatterData) {
-        drupalSettings.OODSPFormatterData[frameId]['width'] = "100%";
-        drupalSettings.OODSPFormatterData[frameId]['height'] = "100%";
-        drupalSettings.OODSPFormatterData[frameId].locale = drupalSettings.OODSP_Settings.locale;
-        if (drupalSettings.OODSP_Settings.isAnonymous) {
-          if (! drupalSettings.OODSPFormatterData[frameId].hasOwnProperty('requestToken')
-                || drupalSettings.OODSPFormatterData[frameId].requestToken.length <= 0) {
-            Drupal.OODSP_Utils.renderError(frameId, {
-              header: drupalSettings.OODSP_Settings.messages.unauthorizedHeader,
-              message: drupalSettings.OODSP_Settings.messages.unauthorizedMessage,
-            });
-            continue;
-          }
+        drupalSettings.OODSPFormatterData[frameId] = Object.assign( drupalSettings.OODSPFormatterData[frameId], defaultConfig );
 
-          DocspaceIntegrationSdk.logout(
-            frameId,
-            function() {
-              DocSpace.SDK.initFrame(drupalSettings.OODSPFormatterData[frameId]);
-            }
-          );
-          continue;
-        }
-
-        DocspaceIntegrationSdk.loginByPasswordHash(
-          frameId,
-          drupalSettings.OODSP_Settings.currentUser,
-          function () {
-            return Drupal.OODSP_Utils.getPasswordHash()
-          },
-          function () {
-            DocSpace.SDK.initFrame(drupalSettings.OODSPFormatterData[frameId]);
-          },
-          function () {
-            if (! drupalSettings.OODSPFormatterData[frameId].hasOwnProperty('requestToken')
-                  || drupalSettings.OODSPFormatterData[frameId].requestToken.length <= 0) {
-              Drupal.OODSP_Utils.renderError(frameId, {
-                header: drupalSettings.OODSP_Settings.messages.unauthorizedHeader,
-                message: drupalSettings.OODSP_Settings.messages.unauthorizedMessage,
-              });
-              return;
-            }
-
+        if (firstElement.frameId == frameId) {
+          if (drupalSettings.OODSP_Settings.isAnonymous) {
             DocspaceIntegrationSdk.logout(
-              frameId,
+              firstElement.frameId,
               function () {
-                DocSpace.SDK.initFrame( config );
+                _initAllFrames(Object.values(drupalSettings.OODSPFormatterData), true);
+              }
+            );
+          } else {
+            DocspaceIntegrationSdk.loginByPasswordHash(
+              firstElement.frameId,
+              drupalSettings.OODSP_Settings.currentUser,
+              function () {
+                return Drupal.OODSP_Utils.getPasswordHash()
+              },
+              function () {
+                _initAllFrames(Object.values(drupalSettings.OODSPFormatterData), false);
+              },
+              function () {
+                DocspaceIntegrationSdk.logout(
+                  firstElement.frameId,
+                  function () {
+                    _initAllFrames(Object.values(drupalSettings.OODSPFormatterData), true);
+                  }
+                );
               }
             );
           }
-        );
+        } else {
+          DocSpace.SDK.initSystem({
+              frameId: frameId,
+              width: "100%",
+              height: "100%",
+              waiting: true
+          });
+        }
       }
     }
   ).catch(
-    function() {
+    function(e) {
       for (var frameId in drupalSettings.OODSPFormatterData) {
         Drupal.OODSP_Utils.renderError(frameId, {message: drupalSettings.OODSP_Settings.messages.docspaceUnavailable});
       }
     }
   );
+
+  const _initAllFrames = (oodspConfigs, requiredRequestToken) => {
+    for (var config of oodspConfigs) {
+      if (requiredRequestToken
+        && (!config.hasOwnProperty('requestToken') || config.requestToken.length <= 0)) {
+
+        if (DocSpace.SDK.frames[config.frameId] != null) {
+          DocSpace.SDK.frames[config.frameId].destroyFrame();
+        }
+
+        Drupal.OODSP_Utils.renderError(config.frameId, {
+          header: drupalSettings.OODSP_Settings.messages.unauthorizedHeader,
+          message: drupalSettings.OODSP_Settings.messages.unauthorizedMessage,
+        });
+
+        continue;
+      }
+
+      DocSpace.SDK.frames[config.frameId].initFrame(config);
+    }
+  }
 
 })(Drupal);