+++ js/onlyoffice-docspace-login.js
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -19,6 +19,8 @@
 (function ($, Drupal) {
   const frameId = 'oodsp-login-frame';
 
+  $loginForm = $('#onlyoffice-docspace-login');
+
   DocspaceIntegrationSdk.initScript('oodsp-api-js', drupalSettings.OODSP_Settings.url)
     .then(function (e) {
       DocSpace.SDK.initSystem(
@@ -28,18 +30,23 @@
             onAppReady: async function () {
               const userInfo = await DocSpace.SDK.frames[frameId].getUserInfo();
               if (userInfo && userInfo.email === drupalSettings.OODSP_Settings.currentUser) {
-                window.location.replace(Url$('input[name="redirect"]').val() || drupalSettings.OODSP_Settings.adminUrl);
+                var redirect = drupalSettings.OODSP_Settings.adminUrl;
+
+                if ($loginForm.find('input[name="redirect"]').length > 0) {
+                  redirect = loginForm.find('input[name="redirect"]').val();
+                }
+
+                window.location.replace(redirect);
               }
             }
           }
         }
       );
     }).catch(function () {
-      window.location.replace($('input[name="redirect"]').val() || drupalSettings.OODSP_Settings.adminUrl);
+      window.location.replace(drupalSettings.OODSP_Settings.adminUrl);
     });
 
   const messages = new Drupal.Message();
-  $loginForm = $('#onlyoffice-docspace-login');
 
   $('input[name="passwordHash"]').val('');
   $loginForm.on('submit', async function (event) {